[Migration 4-5] Feature: rollup/scroll
Closes #3429
