+++ client/features/rollup/scroll/components/ScrollL2TxnBatchDA.tsx
@@ -2,7 +2,7 @@
 
 import React from 'react';
 
-import type { ScrollL2TxnBatch } from 'types/api/scrollL2';
+import type { ScrollL2TxnBatch } from 'client/features/rollup/scroll/types/api';
 import type { ExcludeUndefined } from 'types/utils';
 
 import { Badge } from 'toolkit/chakra/badge';
