+++ client/features/rollup/scroll/pages/batch-details/ScrollL2TxnBatch.tsx
@@ -12,19 +12,21 @@ import { BLOCK } from 'client/slices/block/stubs/block';
 import TxsWithFrontendSorting from 'client/slices/tx/pages/index/list/TxsWithFrontendSorting';
 import { TX } from 'client/slices/tx/stubs/tx';
 
+import { SCROLL_L2_TXN_BATCH } from 'client/features/rollup/scroll/stubs';
+
 import throwOnAbsentParamError from 'client/shared/errors/throw-on-absent-param-error';
 import throwOnResourceLoadError from 'client/shared/errors/throw-on-resource-load-error';
 import useIsMobile from 'client/shared/hooks/useIsMobile';
 import getQueryParamString from 'client/shared/router/get-query-param-string';
 
-import { SCROLL_L2_TXN_BATCH } from 'stubs/scrollL2';
 import { generateListStub } from 'stubs/utils';
 import RoutedTabs from 'toolkit/components/RoutedTabs/RoutedTabs';
 import TextAd from 'ui/shared/ad/TextAd';
 import PageTitle from 'ui/shared/Page/PageTitle';
 import Pagination from 'ui/shared/pagination/Pagination';
 import useQueryWithPages from 'ui/shared/pagination/useQueryWithPages';
-import ScrollL2TxnBatchDetails from 'ui/txnBatches/scrollL2/ScrollL2TxnBatchDetails';
+
+import ScrollL2TxnBatchDetails from './ScrollL2TxnBatchDetails';
 
 const TAB_LIST_PROPS = {
   marginBottom: 0,
