+++ client/features/rollup/scroll/pages/batch-details/ScrollL2TxnBatchDetails.tsx
@@ -4,7 +4,7 @@ import type { UseQueryResult } from '@tanstack/react-query';
 import { useRouter } from 'next/router';
 import React from 'react';
 
-import type { ScrollL2TxnBatch } from 'types/api/scrollL2';
+import type { ScrollL2TxnBatch } from 'client/features/rollup/scroll/types/api';
 
 import { route } from 'nextjs-routes';
 
@@ -12,6 +12,8 @@ import type { ResourceError } from 'client/api/resources';
 
 import BlockEntityL1 from 'client/features/rollup/common/components/BlockEntityL1';
 import TxEntityL1 from 'client/features/rollup/common/components/TxEntityL1';
+import ScrollL2TxnBatchDA from 'client/features/rollup/scroll/components/ScrollL2TxnBatchDA';
+import ScrollL2TxnBatchStatus from 'client/features/rollup/scroll/components/ScrollL2TxnBatchStatus';
 
 import throwOnResourceLoadError from 'client/shared/errors/throw-on-resource-load-error';
 
@@ -19,12 +21,10 @@ import { layerLabels } from 'lib/rollups/utils';
 import { Link } from 'toolkit/chakra/link';
 import { Skeleton } from 'toolkit/chakra/skeleton';
 import isCustomAppError from 'ui/shared/AppError/isCustomAppError';
-import ScrollL2TxnBatchDA from 'ui/shared/batch/ScrollL2TxnBatchDA';
 import DataFetchAlert from 'ui/shared/DataFetchAlert';
 import * as DetailedInfo from 'ui/shared/DetailedInfo/DetailedInfo';
 import DetailedInfoTimestamp from 'ui/shared/DetailedInfo/DetailedInfoTimestamp';
 import PrevNext from 'ui/shared/PrevNext';
-import ScrollL2TxnBatchStatus from 'ui/shared/statusTag/ScrollL2TxnBatchStatus';
 interface Props {
   query: UseQueryResult<ScrollL2TxnBatch, ResourceError>;
 }
