+++ client/features/rollup/scroll/pages/batches/ScrollL2TxnBatches.tsx
@@ -5,7 +5,8 @@ import React from 'react';
 
 import useApiQuery from 'client/api/hooks/useApiQuery';
 
-import { SCROLL_L2_TXN_BATCH } from 'stubs/scrollL2';
+import { SCROLL_L2_TXN_BATCH } from 'client/features/rollup/scroll/stubs';
+
 import { generateListStub } from 'stubs/utils';
 import { Skeleton } from 'toolkit/chakra/skeleton';
 import { ACTION_BAR_HEIGHT_DESKTOP } from 'ui/shared/ActionBar';
@@ -13,8 +14,9 @@ import DataListDisplay from 'ui/shared/DataListDisplay';
 import PageTitle from 'ui/shared/Page/PageTitle';
 import useQueryWithPages from 'ui/shared/pagination/useQueryWithPages';
 import StickyPaginationWithText from 'ui/shared/StickyPaginationWithText';
-import ScrollL2TxnBatchesListItem from 'ui/txnBatches/scrollL2/ScrollL2TxnBatchesListItem';
-import ScrollL2TxnBatchesTable from 'ui/txnBatches/scrollL2/ScrollL2TxnBatchesTable';
+
+import ScrollL2TxnBatchesListItem from './ScrollL2TxnBatchesListItem';
+import ScrollL2TxnBatchesTable from './ScrollL2TxnBatchesTable';
 
 const ScrollL2TxnBatches = () => {
   const { data, isError, isPlaceholderData, pagination } = useQueryWithPages({
