+++ client/features/rollup/scroll/pages/deposits/ScrollL2Deposits.tsx
@@ -5,18 +5,20 @@ import React from 'react';
 
 import useApiQuery from 'client/api/hooks/useApiQuery';
 
+import { SCROLL_L2_MESSAGE_ITEM } from 'client/features/rollup/scroll/stubs';
+
 import { layerLabels } from 'lib/rollups/utils';
-import { SCROLL_L2_MESSAGE_ITEM } from 'stubs/scrollL2';
 import { generateListStub } from 'stubs/utils';
 import { Skeleton } from 'toolkit/chakra/skeleton';
 import { rightLineArrow, nbsp } from 'toolkit/utils/htmlEntities';
-import ScrollL2DepositsListItem from 'ui/deposits/scrollL2/ScrollL2DepositsListItem';
-import ScrollL2DepositsTable from 'ui/deposits/scrollL2/ScrollL2DepositsTable';
 import { ACTION_BAR_HEIGHT_DESKTOP } from 'ui/shared/ActionBar';
 import DataListDisplay from 'ui/shared/DataListDisplay';
 import PageTitle from 'ui/shared/Page/PageTitle';
 import useQueryWithPages from 'ui/shared/pagination/useQueryWithPages';
 import StickyPaginationWithText from 'ui/shared/StickyPaginationWithText';
+
+import ScrollL2DepositsListItem from './ScrollL2DepositsListItem';
+import ScrollL2DepositsTable from './ScrollL2DepositsTable';
 
 const ScrollL2Deposits = () => {
   const { data, isError, isPlaceholderData, pagination } = useQueryWithPages({
