+++ client/features/rollup/scroll/pages/deposits/ScrollL2DepositsTable.tsx
@@ -2,7 +2,7 @@
 
 import React from 'react';
 
-import type { ScrollL2MessageItem } from 'types/api/scrollL2';
+import type { ScrollL2MessageItem } from 'client/features/rollup/scroll/types/api';
 
 import { AddressHighlightProvider } from 'client/slices/address/contexts/address-highlight';
 
@@ -11,7 +11,7 @@ import { layerLabels } from 'lib/rollups/utils';
 import { TableBody, TableColumnHeader, TableHeaderSticky, TableRoot, TableRow } from 'toolkit/chakra/table';
 import TimeFormatToggle from 'ui/shared/time/TimeFormatToggle';
 
-import ScrollL2WithdrawalsTableItem from './ScrollL2WithdrawalsTableItem';
+import ScrollL2DepositsTableItem from './ScrollL2DepositsTableItem';
 
 type Props = {
   items: Array<ScrollL2MessageItem>;
@@ -19,26 +19,26 @@ type Props = {
   isLoading?: boolean;
 };
 
-const ScrollL2WithdrawalsTable = ({ items, top, isLoading }: Props) => {
+const ScrollL2DepositsTable = ({ items, top, isLoading }: Props) => {
   return (
     <AddressHighlightProvider>
       <TableRoot tableLayout="auto" minW="950px">
         <TableHeaderSticky top={ top }>
           <TableRow>
-            <TableColumnHeader>{ layerLabels.current } block</TableColumnHeader>
+            <TableColumnHeader>{ layerLabels.parent } block</TableColumnHeader>
             <TableColumnHeader>Index</TableColumnHeader>
-            <TableColumnHeader>{ layerLabels.current } txn hash</TableColumnHeader>
+            <TableColumnHeader>{ layerLabels.parent } txn hash</TableColumnHeader>
             <TableColumnHeader>
               Timestamp
               <TimeFormatToggle/>
             </TableColumnHeader>
-            <TableColumnHeader>{ layerLabels.parent } txn hash</TableColumnHeader>
-            <TableColumnHeader isNumeric>{ `Value ${ config.chain.currency.symbol }` }</TableColumnHeader>
+            <TableColumnHeader>{ layerLabels.current } txn hash</TableColumnHeader>
+            <TableColumnHeader isNumeric>Value { config.chain.currency.symbol }</TableColumnHeader>
           </TableRow>
         </TableHeaderSticky>
         <TableBody>
           { items.map((item, index) => (
-            <ScrollL2WithdrawalsTableItem key={ String(item.id) + (isLoading ? index : '') } item={ item } isLoading={ isLoading }/>
+            <ScrollL2DepositsTableItem key={ String(item.id) + (isLoading ? index : '') } item={ item } isLoading={ isLoading }/>
           )) }
         </TableBody>
       </TableRoot>
@@ -46,4 +46,4 @@ const ScrollL2WithdrawalsTable = ({ items, top, isLoading }: Props) => {
   );
 };
 
-export default ScrollL2WithdrawalsTable;
+export default ScrollL2DepositsTable;
